back-end-prototype/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alpha-vantage==2.3.1
annotated-types==0.6.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.2.0
flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.0
frozenlist==1.4.0
gotrue==1.3.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.1
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
postgrest==0.13.0
pydantic==2.5.1
pydantic-core==2.14.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
websockets==10.4
werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
